add: args to select mode(1,2)

+++ src/main.ts
@@ -1,53 +1,19 @@
 import './env';
-import { PuppeteerCrawler } from 'crawlee';
-import { getNewVtubers } from './repository/vtubers.repository';
-import { getVideoRouter, newChannelRouter } from './routes.js';
+import { log, PuppeteerCrawler } from 'crawlee';
 import { db } from './database';
+
+import { newVtuberRouter } from './routes/new-vtuber.router';
+import { getDescRouter } from './routes/get-desc.router';
 import { getStreamsToUpdate } from './repository/streams.repository';
+import { getNewVtubers } from './repository/vtubers.repository';
 
-// 1. 신규 V튜버 확인
-// 2. 새로운 라이브 내역 확인
-// 3. 확인 안된 동영상 확인
-
-// const newChannelCrawler = new PuppeteerCrawler({
-//   // proxyConfiguration: new ProxyConfiguration({ proxyUrls: ['...'] }),
-//   requestHandler: newChannelrouter,
-//   // Comment this option to scrape the full website.
-//   maxRequestsPerCrawl: 20,
-//   headless: false,
-//   autoscaledPoolOptions: {
-//     isFinishedFunction: () => disconnectDb(),
-//   },
-//   browserPoolOptions: {
-//     fingerprintOptions: {
-//       fingerprintGeneratorOptions: {
-//         locales: ['ja-JP'],
-//       },
-//     },
-//   },
-// });
-// //Query new channels
-// const newChannels = await getNewVtubers();
-// const newChannelUrls = newChannels.map((data) => data.url);
-//
-// await newChannelCrawler.addRequests(newChannelUrls);
-// await newChannelCrawler.run();
-
-const getStreams = async (): Promise<string[]> => {
-  const streams = await getStreamsToUpdate();
-  return streams.map(
-    (stream) => `https://youtube.com/watch?v=${stream.stream_id}`,
-  );
-};
-
-const getVideoCrawler = new PuppeteerCrawler({
-  // proxyConfiguration: new ProxyConfiguration({ proxyUrls: ['...'] }),
-  requestHandler: getVideoRouter,
-  // Comment this option to scrape the full website.
-  maxRequestsPerCrawl: 20,
-  headless: false,
+// 신규 vTuber 크롤링 런처
+const newVtuberCrawler = new PuppeteerCrawler({
+  requestHandler: newVtuberRouter,
+  launchContext: { launchOptions: { args: ['--lang=ja'] } },
+  headless: true,
   autoscaledPoolOptions: {
-    isFinishedFunction: () => disconnectDb(),
+    isFinishedFunction: () => finishCrawling(),
   },
   browserPoolOptions: {
     fingerprintOptions: {
@@ -58,11 +24,67 @@ const getVideoCrawler = new PuppeteerCrawler({
   },
 });
 
-await getVideoCrawler.addRequests(await getStreams());
+async function finishCrawling() {
+  const newChannels = await getNewVtubers();
+  const newChannelUrls = newChannels.map((data) => data.url);
+  if (newChannelUrls.length == 0) {
+    await db.destroy();
+    return true;
+  }
+  await newVtuberCrawler.addRequests(newChannelUrls);
+  return false;
+}
 
-await getVideoCrawler.run();
+// 라이브 방송 상세 정보 크롤링 런처
+const getDescCrawler = new PuppeteerCrawler({
+  requestHandler: getDescRouter,
+  headless: true,
+  autoscaledPoolOptions: {
+    isFinishedFunction: () => finishGetDesc(),
+  },
+  launchContext: { launchOptions: { args: ['--lang=ja'] } },
+  browserPoolOptions: {
+    fingerprintOptions: {
+      fingerprintGeneratorOptions: {
+        locales: ['ja-JP'],
+      },
+    },
+  },
+});
 
-async function disconnectDb() {
-  await db.destroy();
-  return true;
+const getStreams = async (): Promise<string[]> => {
+  const streams = await getStreamsToUpdate();
+  log.info(`got ${streams.length} of streams`);
+  return streams.map(
+    (stream) => `https://youtube.com/watch?v=${stream.stream_id}`,
+  );
+};
+
+async function finishGetDesc() {
+  const streams = await getStreams();
+  if (streams.length === 0) {
+    await db.destroy();
+    return true;
+  }
+  await getDescCrawler.addRequests(streams);
+  return false;
+}
+
+// 인수 ( mode ) 수신
+const [, , ...args] = process.argv;
+
+// 인수 파싱 로직
+interface Options {
+  mode?: string;
+}
+const options: Options = {};
+for (const arg of args) {
+  const [key, value] = arg.split('=');
+  options[key.replace(/^--/, '')] = value;
+}
+
+if (options.mode === '1') {
+  await newVtuberCrawler.run();
+} else if (options.mode === '2') {
+  await getDescCrawler.run();
 }
